pyproject.toml
[tool.poetry]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = ["Dae-Jeong <v4chelsea@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
gunicorn = "^23.0.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
python-decouple = "^3.8"
requests = "^2.32.3"
django-cors-headers = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
